requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
backcall==0.1.0           # via ipython
click==6.7                # via pip-tools
decorator==4.2.1          # via ipython, traitlets
first==2.0.1              # via pip-tools
gitdb2==2.0.3             # via gitpython
gitpython==2.1.9
ipython-genutils==0.2.0   # via traitlets
ipython==6.3.0
isort==4.3.4
jedi==0.11.1              # via ipython
parso==0.1.1              # via jedi
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==1.11.0
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.5.2         # via pexpect
pygments==2.2.0           # via ipython
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via pip-tools, prompt-toolkit, traitlets
smmap2==2.0.3             # via gitdb2
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
